+++ Marlin/src/pins/gd32f1/pins_VOXELAB_AQUILA.h
@@ -1,6 +1,6 @@
 /**
  * Marlin 3D Printer Firmware
- * Copyright (c) 2022 MarlinFirmware [https://github.com/MarlinFirmware/Marlin]
+ * Copyright (c) 2023 MarlinFirmware [https://github.com/MarlinFirmware/Marlin]
  *
  * Based on Sprinter and grbl.
  * Copyright (c) 2011 Camiel Gubbels / Erik van der Zalm
@@ -27,7 +27,7 @@
  * Uses CREALITY V4 (STM32F103RE / STM32F103RC) board pin assignments
  */
 
-#define BOARD_INFO_NAME      "Aquila v1.0.1"
+#define BOARD_INFO_NAME        "Aquila v1.0.1"
 #ifndef DEFAULT_MACHINE_NAME
   #define DEFAULT_MACHINE_NAME "Aquila"
 #endif
