+++ Marlin/src/pins/hc32f4/pins_AQUILA_101.h
@@ -152,23 +152,23 @@
 #define EXP1_07_PIN                         PB12  // EN2
 #define EXP1_08_PIN                         PB15  // EN1
 
-#if ENABLED(CR10_STOCKDISPLAY) // LCD used for C2
+#if ENABLED(CR10_STOCKDISPLAY)                    // LCD used for C2
 #undef LCD_SERIAL_PORT
-#define LCD_SERIAL_PORT 1
+#define LCD_SERIAL_PORT                        1
 
-  #define LCD_PINS_RS                       EXP1_07_PIN
-  #define LCD_PINS_EN                       EXP1_08_PIN
-  #define LCD_PINS_D4                       EXP1_06_PIN
+  #define LCD_PINS_RS                EXP1_07_PIN
+  #define LCD_PINS_EN                EXP1_08_PIN
+  #define LCD_PINS_D4                EXP1_06_PIN
 
-  #define BTN_ENC                           EXP1_02_PIN
-  #define BTN_EN1                           EXP1_03_PIN
-  #define BTN_EN2                           EXP1_05_PIN
+  #define BTN_ENC                    EXP1_02_PIN
+  #define BTN_EN1                    EXP1_03_PIN
+  #define BTN_EN2                    EXP1_05_PIN
 
   #ifndef HAS_PIN_27_BOARD
-    #define BEEPER_PIN                      EXP1_01_PIN
+    #define BEEPER_PIN               EXP1_01_PIN
   #endif
 
-#elif ANY(HAS_DWIN_E3V2, IS_DWIN_MARLINUI) // LCD used for X2
+#elif ANY(HAS_DWIN_E3V2, IS_DWIN_MARLINUI)        // LCD used for X2
   /**
    * Display pinout (display side, so RX/TX are swapped)
    *
