+++ buildroot/share/sublime/MarlinFirmware.sublime-project
@@ -30,6 +30,7 @@
 		"ensure_newline_at_eof_on_save": true,
 		"tab_size": 2,
 		"translate_tabs_to_spaces": true,
-		"trim_trailing_white_space_on_save": true
+		"trim_trailing_white_space_on_save": true,
+		"uncrustify_config" : "${project_dir}/../extras/uncrustify.cfg"
 	}
 }
